requirements.txt
aiofiles==23.2.1
aiogoogle==5.6.0
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
black==23.12.1
cachetools==5.3.2
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
Django==5.0.1
django-social-auth==0.7.28
djangorestframework==3.14.0
frozenlist==1.4.1
google-auth==2.27.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
idna==3.6
multidict==6.0.5
mypy-extensions==1.0.0
oauth2==1.9.0.post1
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dotenv==1.0.1
python-openid==2.2.5
python-telegram-bot==20.7
pytz==2023.3.post1
rsa==4.9
sniffio==1.3.0
sqlparse==0.4.4
tonyg-rfc3339==0.1
tzdata==2023.4
yarl==1.9.4
